% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_CAplot.R
\name{DESeq2_CAplot}
\alias{DESeq2_CAplot}
\title{Produces CA plot}
\usage{
DESeq2_CAplot(feat_count, DESeq2_object, outdir, feature_name)
}
\arguments{
\item{feat_count}{Count table from featurecounts}

\item{DESeq2_object}{A DESeq2 object}

\item{outdir}{A folder to write all the files}

\item{feature_name}{name of the feature from gff table that is being counted}
}
\value{
creates PCA plot of PDF and PNG figures
}
\description{
Produces CA plot
}
