% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_histogram.R
\name{DESeq2_hist}
\alias{DESeq2_hist}
\title{Produces heatmap plots based on the FPKM table}
\usage{
DESeq2_hist(f_table, exp_desn, outdir, score_type, feature_name)
}
\arguments{
\item{f_table}{A csv table with fpm, generated using DESeq2FPKM}

\item{exp_desn}{An experimental design file of PiReT}

\item{outdir}{A folder to write all the files}

\item{score_type}{FPM or FPKM}
}
\value{
creates PDF and PNG figures from FPM/FPKM tables
}
\description{
Produces heatmap plots based on the FPKM table
}
