% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_histogram.R
\name{DESeq2_histogram}
\alias{DESeq2_histogram}
\title{Produces histogram plots based on the FPKM or FPM table}
\usage{
DESeq2_histogram(f_table, exp_desn, outdir, score_type, feature_name)
}
\arguments{
\item{f_table}{A csv table with fpm or fpkm, generated using DESeq2FPKM}

\item{exp_desn}{An experimental design file of PiReT}

\item{outdir}{A folder to write all the files}

\item{score_type}{FPM or FPKM}

\item{feature_name}{name of the feature from gff table that is being counted}
}
\value{
creates PDF and PNG figures from FPM/FPKM tables
}
\description{
Produces histogram plots based on the FPKM or FPM table
}
