% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_summary.R
\name{DESeq2_summary}
\alias{DESeq2_summary}
\title{Produces CA plot}
\usage{
DESeq2_summary(object, alpha, pair1, pair2, feature_name, outdir)
}
\arguments{
\item{object}{A DESeq2 object}

\item{alpha}{alpha}

\item{pair1}{condition 1}

\item{pair2}{condition 2}

\item{feature_name}{name of the feature that is being compared, for example, CDS}

\item{outdir}{A folder to write all the files}
}
\value{
creates a csv file with summary information from deseq2 object with number of genes that up, down, and not significantly different between two conditions.
}
\description{
Produces CA plot
}
