% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_violin.R
\name{DESeq2_violin}
\alias{DESeq2_violin}
\title{Produces violin plots based on the FPKM/FPM table}
\usage{
DESeq2_violin(f_table, exp_desn, outdir, score_type, feature_name)
}
\arguments{
\item{f_table}{A csv table with fpm or fpkm, generated using DESeq2FPKM or DESeq2FPM}

\item{exp_desn}{An experimental design file of PiReT}

\item{outdir}{A folder to write all the files}

\item{score_type}{FPM or FPKM}

\item{feature_name}{name of the feature from gff table that is being counted}
}
\value{
creates PDF and PNG figures from FPM/FPKM tables
}
\description{
Produces violin plots based on the FPKM/FPM table
}
