% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2FPKM.R
\name{DESeq2FPKM}
\alias{DESeq2FPKM}
\title{Calculates FPKM from DESeq2 object and creates corresponding figures}
\usage{
DESeq2FPKM(DESeq2_object, feat_count, outdir)
}
\arguments{
\item{DESeq2_object}{created using feat2deseq2 function}

\item{feat_count}{A tsv file from featurecount}

\item{outdir}{A folder to write all the files}
}
\value{
creates CSV fpkm table base off DEDeq2 object
}
\description{
Calculates FPKM from DESeq2 object and creates corresponding figures
}
