% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2FPM.R
\name{DESeq2FPM}
\alias{DESeq2FPM}
\title{Calculates FPM from DESeq2 object}
\usage{
DESeq2FPM(DESeq2_object, feat_count, outdir)
}
\arguments{
\item{DESeq2_object}{created using feat2deseq2 function}

\item{feat_count}{A tsv file from featurecount}

\item{outdir}{A folder to write all the files}
}
\value{
creates CSV fpm table based off DEDeq2 object
}
\description{
Calculates FPM from DESeq2 object
}
