% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat2deseq2.R
\name{feat2deseq2}
\alias{feat2deseq2}
\title{Creates a DESeq2 object from featurecount table and experimental design file}
\usage{
feat2deseq2(feat_count, exp_desn)
}
\arguments{
\item{feat_count}{A tsv file from featurecount}

\item{exp_desn}{experimental design file for PiReT}
}
\value{
DESeq2 object with features that dont have any counts removed
}
\description{
Creates a DESeq2 object from featurecount table and experimental design file
}
