% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPKM_heatmap.R
\name{FPKM_heatmap}
\alias{FPKM_heatmap}
\title{Produces heatmap plots based on the FPKM table}
\usage{
FPKM_heatmap(fpkm_table, exp_desn, outdir)
}
\arguments{
\item{fpkm_table}{A csv table with fpkm, generated using DESeq2FPKM}

\item{exp_desn}{An experimental design file of PiReT}

\item{outdir}{A folder to write all the files}
}
\value{
creates PDF and PNG figures from FPKM tables
}
\description{
Produces heatmap plots based on the FPKM table
}
