% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPM_heatmap.R
\name{FPM_heatmap}
\alias{FPM_heatmap}
\title{Produces heatmap plots based on the FPKM table}
\usage{
FPM_heatmap(fpm_table, exp_desn, outdir)
}
\arguments{
\item{fpm_table}{A csv table with fpm, generated using DESeq2FPKM}

\item{exp_desn}{An experimental design file of PiReT}

\item{outdir}{A folder to write all the files}
}
\value{
creates PDF and PNG figures from FPM tables
}
\description{
Produces heatmap plots based on the FPKM table
}
