% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage_analysis.R
\name{gage_analysis}
\alias{gage_analysis}
\title{Conducts GAGE analysis using fold change data.}
\usage{
gage_analysis(fc, type, org, out_dir)
}
\arguments{
\item{fc}{A table produced by edgeR or DESeq2 that contains fold change information.}

\item{type}{edgeR or DESeq2.}

\item{org}{An organism code provided by KEGG.}

\item{out_dir}{Output directory where tables are written.}
}
\value{
Table with list of pathways and their corresponding p-values indicating if they are differently expressed between two conditions
}
\description{
Conducts GAGE analysis using fold change data.
}
