% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_pathway.R
\name{map_pathway}
\alias{map_pathway}
\title{Gets pathway maps and paints colors based on expression level of genes}
\usage{
map_pathway(sig_table, type, org_code, out)
}
\arguments{
\item{sig_table}{A table file from DESeq2 or edgeR}

\item{type}{Is it DESeq2 or edgeR?}

\item{out}{output folder}

\item{org}{KEGG organism code}
}
\value{
A folder containing pathway with fold change value mapped
}
\description{
Gets pathway maps and paints colors based on expression level of genes
}
